+++ PowerShellTools/DebugEngine/ScriptDebugger.cs
@@ -110,87 +110,6 @@ namespace PowerShellTools.DebugEngine
         }
 
         /// <summary>
-        /// Unused at the moment. Will be used for remote debugging scripts.
-        /// </summary>
-        /// <param name="remoteRunspace"></param>
-        public void RegisterRemoteFileOpenEvent(Runspace remoteRunspace)
-        {
-            remoteRunspace.Events.ReceivedEvents.PSEventReceived += new PSEventReceivedEventHandler(this.HandleRemoteSessionForwardedEvent);
-            if (remoteRunspace.RunspaceStateInfo.State != RunspaceState.Opened || remoteRunspace.RunspaceAvailability != RunspaceAvailability.Available)
-            {
-                return;
-            }
-            using (PowerShell powerShell = PowerShell.Create())
-            {
-                powerShell.Runspace = remoteRunspace;
-                powerShell.AddScript("\r\n            param (\r\n                [string] $PSEditFunction\r\n            )\r\n\r\n            Register-EngineEvent -SourceIdentifier PSISERemoteSessionOpenFile -Forward\r\n\r\n            if ((Test-Path -Path 'function:\\global:PSEdit') -eq $false)\r\n            {\r\n                Set-Item -Path 'function:\\global:PSEdit' -Value $PSEditFunction\r\n            }\r\n        ").AddParameter("PSEditFunction", "\r\n            param (\r\n                [Parameter(Mandatory=$true)] [String[]] $FileNames\r\n            )\r\n\r\n            foreach ($fileName in $FileNames)\r\n            {\r\n                dir $fileName | where { ! $_.PSIsContainer } | foreach {\r\n                    $filePathName = $_.FullName\r\n\r\n                    # Get file contents\r\n                    $contentBytes = Get-Content -Path $filePathName -Raw -Encoding Byte\r\n\r\n                    # Notify client for file open.\r\n                    New-Event -SourceIdentifier PSISERemoteSessionOpenFile -EventArguments @($filePathName, $contentBytes) > $null\r\n                }\r\n            }\r\n        ");
-                try
-                {
-                    powerShell.Invoke();
-                }
-                catch (RemoteException)
-                {
-                }
-            }
-        }
-
-        /// <summary>
-        /// Unused at the moment. Will be used for remote debugging scripts.
-        /// </summary>
-        /// <param name="remoteRunspace"></param>
-        public void UnregisterRemoteFileOpenEvent(Runspace remoteRunspace)
-        {
-            remoteRunspace.Events.ReceivedEvents.PSEventReceived -= new PSEventReceivedEventHandler(this.HandleRemoteSessionForwardedEvent);
-            if (remoteRunspace.RunspaceStateInfo.State != RunspaceState.Opened || remoteRunspace.RunspaceAvailability != RunspaceAvailability.Available)
-            {
-                return;
-            }
-            using (PowerShell powerShell = PowerShell.Create())
-            {
-                powerShell.Runspace = remoteRunspace;
-                powerShell.AddScript("\r\n            if ((Test-Path -Path 'function:\\global:PSEdit') -eq $true)\r\n            {\r\n                Remove-Item -Path 'function:\\global:PSEdit' -Force\r\n            }\r\n\r\n            Get-EventSubscriber -SourceIdentifier PSISERemoteSessionOpenFile -EA Ignore | Remove-Event\r\n        ");
-                try
-                {
-                    powerShell.Invoke();
-                }
-                catch (RemoteException)
-                {
-                }
-            }
-        }
-
-        /// <summary>
-        /// Unused at the moment. Will be used for remote debugging scripts.
-        /// </summary>
-        /// <param name="sender"></param>
-        /// <param name="args"></param>
-        private void HandleRemoteSessionForwardedEvent(object sender, PSEventArgs args)
-        {
-            if (args.SourceIdentifier.Equals("PSISERemoteSessionOpenFile", StringComparison.OrdinalIgnoreCase))
-            {
-                string text = null;
-                byte[] array = null;
-                try
-                {
-                    if (args.SourceArgs.Length == 2)
-                    {
-                        text = (args.SourceArgs[0] as string);
-                        array = (byte[])(args.SourceArgs[1] as PSObject).BaseObject;
-                    }
-                    if (!string.IsNullOrEmpty(text) && array != null)
-                    {
-                        // bool flag;
-                        // this.LoadFile(text, array, out flag);
-                    }
-                }
-                catch
-                {
-
-                }
-            }
-        }
-
-        /// <summary>
         /// Clears existing breakpoints for the current runspace.
         /// </summary>
         private void ClearBreakpoints()
